Audio working
GOOD LORD.. I've been trying to fix this for days and days and it turned out it was a SHIELDING issue! Chalk it up to my absolute inexperience with hardware. Turns out the servo is electrically noisy. When I unplugged it, the init and "Skit - names" audio both played perfectly ten times in row. When the servo was plugged in the best I ever saw was playing the init and 30% of "Skit - names" before crashing.

So now I'm going to try shielding the SD card and its wires.